maquinas/requirements.txt
numpy
django
psycopg2-binary
pandas
gymnasium
stable-baselines3
websocket-client
tensorboard
django-channels
channels-redis
daphne
 
